Revert "Async callbacks for processing GetColumnProfile RPC requests (#473)"
This reverts commit 9f864b3b9575aa886259840e755c5b401f39a3cb.
