+++ crates/ark/src/data_explorer/mod.rs
@@ -5,11 +5,9 @@
 //
 //
 
-pub mod column_profile;
 pub mod export_selection;
 pub mod format;
 pub mod histogram;
 pub mod r_data_explorer;
 pub mod summary_stats;
-pub mod table;
 pub mod utils;
